examples simplified
