+++ docs/user_guide/slurm.md
@@ -84,7 +84,7 @@ srun examples/slurm/launch.sh
 
 ```bash
 # launch.sh (runs once per node)
-exec traceml run examples/ddp_minimal.py \
+exec traceml run examples/distributed/ddp_minimal.py \
   --mode=summary \
   --run-name="${RUN_NAME}" \
   --nnodes="${SLURM_NNODES}" \
@@ -143,7 +143,7 @@ traceml run train.py ... --aggregator-bind-host=10.0.0.5
 A single node's launch (what the wrapper runs) looks like this:
 
 ```bash
-traceml run examples/ddp_minimal.py \
+traceml run examples/distributed/ddp_minimal.py \
   --mode=summary \
   --run-name=ddp-demo \
   --nnodes=2 \
@@ -153,7 +153,7 @@ traceml run examples/ddp_minimal.py \
   --master-port=29500
 ```
 
-`examples/ddp_minimal.py` is a runnable DDP script that already calls
+`examples/distributed/ddp_minimal.py` is a runnable DDP script that already calls
 `traceml.init(...)` and `traceml.trace_step(...)`.
 
 ## Full template
